% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Data1}
\alias{Data1}
\title{Data1}
\format{A data frame with 200 observations on the following 2 variables.
\describe{ \item{list("V1")}{a numeric vector} \item{list("V2")}{a numeric
vector} }}
\usage{
Data1
}
\description{
Data1 dataset used to demonstrate the use of SNFtool.
}
\examples{
data(Data1)
}
\keyword{datasets}

