% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Data2}
\alias{Data2}
\title{Data2}
\format{A data frame with 200 observations on the following 2 variables.
\describe{ \item{list("V3")}{a numeric vector} \item{list("V4")}{a numeric
vector} }}
\usage{
Data2
}
\description{
Data2 dataset used to demonstrate the use of SNFtool.
}
\examples{
data(Data2)
}
\keyword{datasets}

