\name{dataL}
\alias{dataL}
\docType{data}
\title{
dataL
}
\description{
Dataset used to provide an example of predicting the new labels with label propagation.
}
\usage{
data(dataL)
}
\format{
  The format is:
List of 2
 $ : num [1:600, 1:76] 0.0659 0.0491 0.0342 0.0623 0.062 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:600] "V1" "V2" "V3" "V4" ...
  .. ..$ : NULL
 $ : int [1:600, 1:240] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:600] "V1" "V2" "V3" "V4" ...
  .. ..$ : NULL
}
\examples{
data(dataL)
}
\keyword{datasets}
