% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{label}
\alias{label}
\title{Labels for dataL dataset}
\format{The format is: int [1:600] 1 1 1 1 1 1 1 1 1 1 ...}
\usage{
label
}
\description{
The ground truth for dataL dataset
}
\examples{
data(label)
}
\keyword{datasets}

