% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/robustSimilarity.R
\name{quantize}
\alias{quantize}
\title{Quantize values}
\usage{
quantize(x, quantiles = 20)
}
\arguments{
\item{x}{numeric vector to quantize}

\item{quantiles}{number of quantiles to use}
}
\value{
a numeric vector of the same length as x, giving the quantiles of values in x
}
\description{
Replaces values by their quantile number.
}
\examples{
quantize(runif(100))

# ranking
quantize(runif(100), 100)

# quantization can deal with high kurtosis data
quantize(c(rnorm(100,-1000), rnorm(100,1000)))
}

