% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/standardNormalization.R
\name{standardNormalization}
\alias{standardNormalization}
\title{Standard Normalization}
\usage{
standardNormalization(x)
}
\arguments{
\item{x}{The unnormalized data.}
}
\value{
The data normalized.
}
\description{
Normalize each column of the input data to have mean 0 and standard
deviation 1.
}
\examples{
## Data1 is of size n x d_1,
## where n is the number of patients, d_1 is the number of genes,
## Data2 is of size n x d_2,
## where n is the number of patients, d_2 is the number of methylation
data(Data1)
data(Data2)

Data1 = standardNormalization(Data1);
Data2 = standardNormalization(Data2);
}
\author{
Dr. Anna Goldenberg, Bo Wang, Aziz Mezlini, Feyyaz Demir
}

