\name{Data1}
\alias{Data1}
\docType{data}
\title{
Data1
}
\description{
Data1 dataset used to demonstrate the use of SNFtool. 
}
\usage{
data(Data1)
}
\format{
A data frame with 200 observations on the following 2 variables.
  \describe{
    \item{\code{V1}}{a numeric vector}
    \item{\code{V2}}{a numeric vector}
  }
}
\examples{
data(Data1)
}
\keyword{datasets}
