\name{Data2}
\alias{Data2}
\docType{data}
\title{
Data2
}
\description{
Data2 dataset used to demonstrate the use of SNFtool. 
}
\usage{
data(Data2)
}
\format{
A data frame with 200 observations on the following 2 variables.
  \describe{
    \item{\code{V3}}{a numeric vector}
    \item{\code{V4}}{a numeric vector}
  }
}
\examples{
data(Data2)
}
\keyword{datasets}
